+++ src/main/java/me/myogoo/extendedterminal/api/integration/itemList/IGuiSlotToIngredientMap.java
@@ -1,0 +1,13 @@
+package me.myogoo.extendedterminal.api.integration.itemList;
+
+import me.myogoo.extendedterminal.api.adapter.recipe.table.ITableRecipeAdapter;
+import me.myogoo.extendedterminal.menu.ETTerminalBaseMenu;
+import net.minecraft.world.item.crafting.Ingredient;
+
+import java.util.Map;
+
+public interface IGuiSlotToIngredientMap {
+    Map<Integer, Ingredient> jei(ETTerminalBaseMenu<?> menu, ITableRecipeAdapter recipe);
+    Map<Integer, Ingredient> emi(ETTerminalBaseMenu<?> menu, ITableRecipeAdapter recipe);
+
+}