+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## Bug Fixes
-- Fix EMI recipe transfer not working for recipes without recipeId
-- Like the processor in ae2, the inscriber no longer consumes the compat press. [@CynSolveroftheAbsolute](https://github.com/CynSolveroftheAbsolute) ([#24](https://github.com/mc-myo-s-mod/ExtendedTerminal/issues/24#issue-3909100480))
+- Fix EMI/JEI recipe gui slot mismatch
+- Refactor Extended Terminal Anvil Logic